# Conflicts:
#	newrelic-security-agent/src/main/java/com/newrelic/agent/security/AgentConfig.java
#	newrelic-security-agent/src/main/java/com/newrelic/agent/security/instrumentator/utils/InstrumentationUtils.java
#	newrelic-security-agent/src/main/java/com/newrelic/agent/security/intcodeagent/controlcommand/ControlCommandProcessor.java
#	newrelic-security-agent/src/main/java/com/newrelic/agent/security/intcodeagent/websocket/EventSendPool.java
#	newrelic-security-agent/src/main/java/com/newrelic/api/agent/security/Agent.java
#	newrelic-security-api/src/main/java/com/newrelic/api/agent/security/schema/HttpRequest.java

+++ newrelic-security-agent/src/main/java/com/newrelic/agent/security/instrumentator/utils/InstrumentationUtils.java
@@ -15,6 +15,7 @@ import com.newrelic.agent.security.intcodeagent.websocket.WSClient;
 import com.newrelic.agent.security.intcodeagent.websocket.WSReconnectionST;
 import org.apache.commons.io.FileUtils;
 import org.java_websocket.framing.CloseFrame;
+import org.apache.commons.lang3.StringUtils;
 
 import java.io.File;
 import java.util.concurrent.TimeUnit;
@@ -76,7 +77,9 @@ public class InstrumentationUtils {
             WSReconnectionST.shutDownPool();
             WSClient.shutDownWSClient(true, CloseFrame.NORMAL, "IAST agent shutting down");
             FileCleaner.cancelTask();
-            FileUtils.deleteQuietly(new File(OsVariablesInstance.getInstance().getOsVariables().getTmpDirectory()));
+            if(StringUtils.isNotBlank(OsVariablesInstance.getInstance().getOsVariables().getTmpDirectory())) {
+                FileUtils.deleteQuietly(new File(OsVariablesInstance.getInstance().getOsVariables().getTmpDirectory()));
+            }
 
         } catch (Throwable e) {
             logger.log(LogLevel.SEVERE, "Error while shutting down executor pools : ", e,
